refactor: name the boot steps and split the settings save
src-tauri/src/boot.rs takes the 280-line setup closure. main now lists
eight steps: build_main_window, disable_webview_autofill, telemetry,
install_close_handler, wire_deep_links, spawn_snapshot_upgrade,
spawn_boot_failsafe. same order, same calls.

spawn_window_size_save turns three levels of nesting into three
guards. main 32 CCN, off the list.

Settings.svelte persistCurrentState 27 CCN threaded a hadError flag
through five saves. commitPinCode, saveApiKey, savePlatformPaths and
lastSavedPlatformPaths each return whether they landed. also off.

not launched in the real app: the boot path moved, nothing drove it.

cargo test --workspace 411, vitest 175, clippy and svelte-check clean.
